区域批量删除方法加上强制删除标识

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/businness/impl/AreaBusinessServiceImpl.java
@@ -3,10 +3,13 @@ package com.deepexi.channel.businness.impl;
 import com.deepexi.channel.businness.AreaBusinessService;
 import com.deepexi.channel.businness.StoreAreaBusinessService;
 import com.deepexi.channel.domain.area.*;
+import com.deepexi.channel.domain.distributor.DistributorAreaRelationDTO;
 import com.deepexi.channel.domain.store.StoreAreaDTO;
 import com.deepexi.channel.domain.store.StoreAreaQuery;
+import com.deepexi.channel.enums.ForceDeleteEnum;
 import com.deepexi.channel.service.AreaService;
 import com.deepexi.channel.service.AreaTypeService;
+import com.deepexi.channel.service.DistributorAreaRelationService;
 import com.deepexi.channel.service.StoreAreaService;
 import com.deepexi.util.CollectionUtil;
 import com.deepexi.util.StringUtil;
@@ -36,6 +39,9 @@ public class AreaBusinessServiceImpl implements AreaBusinessService {
 
     @Autowired
     StoreAreaService storeAreaService;
+
+    @Autowired
+    DistributorAreaRelationService distributorAreaRelationService;
 
     @Transient
     @Override
@@ -255,24 +261,78 @@ public class AreaBusinessServiceImpl implements AreaBusinessService {
 
     @Transient
     @Override
-    public boolean deleteBatch(List<Long> ids) {
+    public boolean deleteBatchByIds(List<Long> ids,Integer forceDelete) {
 
         if(CollectionUtils.isEmpty(ids)){
             return false;
         }
 
-        //删除子节点关联
-        deleteChildren(ids);
+        if(forceDelete== ForceDeleteEnum.NO.getCode()){
 
-        //经销商也有区域关联
+            validateHasChildren(ids);
 
-        //删除门店关联
-        deleteStore(ids);
+            validateHasDistributors(ids);
+
+            validateHasStores(ids);
+
+        }else if(forceDelete==ForceDeleteEnum.YES.getCode()){
+
+            deleteChildren(ids);
+
+            deleteDistributors(ids);
+
+            deleteStores(ids);
+        }
 
         return areaService.deleteBatch(ids);
     }
 
-    private boolean deleteChildren(List<Long> idList){
+    @Override
+    public void validateHasDistributors(List<Long> areaIdList){
+
+        log.info("查询是否有经销商关联");
+
+        List<DistributorAreaRelationDTO> darList = distributorAreaRelationService.findAllByAreaIds(areaIdList);
+
+        if(CollectionUtils.isNotEmpty(darList)){
+
+            throw new ApplicationException("此区域已有经销商关联,不能删除!请解除关联关系后再操作");
+
+        }
+    }
+
+    @Override
+    public boolean deleteDistributors(List<Long> areaIdList){
+
+        log.info("删除经销商关联");
+
+        List<DistributorAreaRelationDTO> darList = distributorAreaRelationService.findAllByAreaIds(areaIdList);
+
+        if(CollectionUtils.isNotEmpty(darList)){
+
+            distributorAreaRelationService.deleteBatchByAreaIds(areaIdList);
+        }
+
+        return Boolean.TRUE;
+    }
+
+    @Override
+    public void validateHasChildren(List<Long> idList){
+
+        log.info("查询是否有子区域关联");
+
+        List<AreaDTO> children = getChildrenByIds(idList);
+
+        if(CollectionUtils.isNotEmpty(children)){
+
+            throw new ApplicationException("已有子区域关联,不能删除!请解除关联后再操作");
+        }
+    }
+
+    @Override
+    public boolean deleteChildren(List<Long> idList){
+
+        log.info("删除子区域关联");
 
         if(CollectionUtils.isNotEmpty(idList)){
 
@@ -284,7 +344,27 @@ public class AreaBusinessServiceImpl implements AreaBusinessService {
         return true;
     }
 
-    private boolean deleteStore(List<Long> ids){
+    @Override
+    public void validateHasStores(List<Long> ids){
+
+        log.info("查询是否有门店关联");
+
+        StoreAreaQuery query = new StoreAreaQuery();
+
+        query.setAreaIds(ids);
+
+        List<StoreAreaDTO> pageList = storeAreaService.findList(query);
+
+        if(CollectionUtils.isNotEmpty(pageList)){
+
+            throw new ApplicationException("此区域已有门店关联,不能删除!请解除所有关联后再操作");
+        }
+    }
+
+    @Override
+    public boolean deleteStores(List<Long> ids){
+
+        log.info("删除门店关联");
 
         StoreAreaQuery query = new StoreAreaQuery();
 
@@ -317,23 +397,13 @@ public class AreaBusinessServiceImpl implements AreaBusinessService {
 
             pageList.forEach(area->{
 
-                if(!parentIdList.contains(area.getId())){
+                if(parentIdList.contains(area.getId())){
 
                     children.add(area);
                 }
             });
         }
         return  children;
-    }
-
-    @Override
-    public boolean deleteById(Long id) {
-
-        log.info("单个删除区域");
-
-        deleteChild(id);
-
-        return areaService.deleteById(id);
     }
 
     private boolean deleteChild(Long parentId){
@@ -356,7 +426,7 @@ public class AreaBusinessServiceImpl implements AreaBusinessService {
                 if(child.getParentId()==parentId){
 
                     //父级节点被删除
-                    child.setParentId(-1L);
+                    child.setParentId(0L);
                 }
             });
         }
